Image-Classifier/requirements.txt
Flask
keras
pillow
tensorflow
numpy
